Fix home dir expansion in setup-repo
- The hooks directory was meant to be created in $HOME/.git/hooks
but that was not working in all bash environments because of the
use of ~ and not $HOME. ~ was not being expanded.

- Changed to '.git/hooks' which is inside the project dir. If we
want to setup the repo dir, that makes no sense to me to use the
global .git dir. People may not want space commander in all obj-c
projects and for sure they dont want it in non obj-c projects.